# server/apidocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PetAdobt API",
    "description": "PetAdobt APIs List",
    "version": "1.0.0"
  },
...
  "paths": {
    "/pets": {
      "get": {
        "tags": ["Pet"],
        "summary": "Get Data Pets",
        "description": "",
        "responses": {
          "200": {
            "description": "Succeed!"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/petadobts": {
      "get": {
        "tags": ["PetAdobt"],
        "summary": "Get Data PetAdobts",
        "description": "",
        "responses": {
          "200": {
            "description": "Succeed!"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}